Extract expansion state into separate class.
